Hierarchical-Retrieval-Augmented-Generation-H-RAG-System-main/requirements.txt
streamlit==1.25.0
PyPDF2==3.0.1
faiss-cpu==1.7.3
numpy==1.24.3
ollama==0.0.1

